Simplify mpl.testing._copy_metadata.
`mpl.testing._copy_metadata` was basically doing the job
of `functools.update_wrapper`.  We don't need to specify
`compat_co_firstlineono` either as pytest will look up
the lineno in the wrapped function
(https://github.com/pytest-dev/pytest/blob/master/_pytest/compat.py#L188).
It's also for this reason that we need to make sure
`__wrapped__` is properly set, even on Py2.

Also remove a useless check on `image_comparison` (if we want an
argument to be compulsory then just mark it as such, this does not
prevent it to be passed as keyword).