FIX: always remove the temporary file in eps print_figure
In cases were a filename is passed in we clean up the temporary file
by renaming it to the output file, however if a file handle is passed
in we copy the contents into that the file and do not clean up the
temporary file.

Use a try...finally block to always remove the temporary file, even if
things go wrong.