Fix testing with tests.py on Py3.6.
Invoking `tests.py` used to fail on Py3.6 due to the use of deprecated
invalid escapes both in the codebase and in third-party code (dateutil).
Fix the codebase, skip the checks on dateutil.

It is strange (and should be worth investigating...) that the CI did not
fail before, as it also uses `tests.py` on Py3.6...