TST: Make image_comparison more pytest-y.
Instead of a heavy do-it-all class, split ImageComparisonDecorator into
a smaller class that does just the comparison stuff and one that does
only nose.

For pytest, use a wrapper function that's decorated only by pytest
decorators, and don't try to modify the function signature. By using a
separate fixture, we can indirectly return the parameterized arguments
instead. This stops pytest from getting confused about what takes what
argument.

The biggest benefit is that test code is now run as the *test*, whereas
previously, it was run as the *setup* causing it to have all sorts of
semantic irregularities.