LogLocator: fix problems with minor ticks. Closes #8111.
The subsetting of minor ticks that was introduced in v2.0
made it hard to know what number a given unlabeled minor
tick represents, so we have decided that minor log ticks
should be all integer multiples of the decade (except the
decade itself).

When major ticks skip decades, then minor ticks are omitted
entirely.