check the type of the output animation filename
Currently it is difficult to determine the error when the animation movie
filename is something other than a string. Even with the debug output of
matplotlib, it is not easy to determin the reason why the subprocess call
fails.

I added a check that explicitly checks the type of the output file variable.