BUG: handle empty levels array in contour, closes #7486
This relates to line contours only, not to filled contours.

This handles the case where an empty array is provided as the
"levels" kwarg, or where for any reason there are no levels within
the data range.  An example is the case of uniform z. The
levels array is then set to the minimum value of z so that
the plotting can proceed, and a warning is generated.