Simplify the PySide QImage memleak fix.
Instead of manipulating refcounts, hide the image buffer behind a numpy
array that does not own the data; this prevents QImage from setting a
positive refcount on the Python buffer that actually is in charge of
deallocating the data upon GC.

The asserts check that the fix actually works.  Replacing the body of
`_make_non_owning_array` by `return buf` make the assert fail.