Warn on freetype missing glyphs.
There is an additional call site to FT_Get_Char_Index in
PyFT2Font_get_char_index but that value is directly returned to Python
so it would be easier to check for the value there and handle it
accordingly (plus, in practice, the same warning will be triggered at
the other two call sites at some point anyways).  But I can replace that
by a warning too.