ENH: improve contourf color autoscaling
Contour level auto-selection: clip unused Locator output:

Locator.tick_values() returns levels beyond the data limits.
In the case of LogLocator with a small data range, the overrun
can be large because it is expanding to the decade points.
In addition, no account was being taken of the "extend" kwarg.
With this changeset, the outermost levels will be the miminum
required to include the data interval in the default case, and
will be reduced when "extend" is used so that some of the data
range will fall in the extended sections.  This is expected to
be rare, however; normally the "extend" kwarg would be used only
when levels are explicitly set, not auto-selected with a Locator.

The contour_hatching test results were modified.  The difference
was reduced by putting the hatches in the same order as they were in the
reference figure, but the images still don't match because
the gray shades have changed, as expected.

(Note that in the original and new reference figures
the pdf and svg backends handle the alpha differently than
the agg backend.  This is unrelated to the PR.)

For logscale contour colors, interpolate in screen space
by taking the geometric mean of adjacent levels.

The test_contour.py::test_contourf_log_extension code and
image have been modified for this change in colors, and to
make the behavior more clear in the case where levels are
not specified.

Add a note to next_api_changes.

Treat contour levels (almost) the same as contourf levels

Update additional test images affected by the contour color changes.

mplot3d: don't assume every contour level has a line.