Contour level auto-selection: clip unused Locator output
Locator.tick_values() returns levels beyond the data limits.
In the case of LogLocator with a small data range, the overrun
can be large because it is expanding to the decade points.
In addition, no account was being taken of the "extend" kwarg.
With this changeset, the outermost levels will be the miminum
required to include the data interval in the default case, and
will be reduced when "extend" is used so that some of the data
range will fall in the extended sections.  This is expected to
be rare, however; normally the "extend" kwarg would be used only
when levels are explicitly set, not auto-selected with a Locator.