partial fix for contour_hatching test failure
This puts the hatches in the same order as they were in the
reference figure, but the images still don't match because
the gray shades are not the same as they were.

Furthermore, looking at the original reference figures shows
that the pdf and svg backends handle the alpha differently than
the agg backend.  Not sure what to do about that.