FIX: properly mix blitting + redraw in Qt4Agg/Qt5Agg
If a draw is requested on the next paint event update the whole
canvas, not just the blit boxes.  Any un-shown blitted updates from
before the full redraw are lost (due to the full update) and any
blitted updates from after the full render (but before the paint
event) will be included (because the renderer updates in-place).

Closes #7366