Revert "Merge pull request #5754 from blink1073/ipython-widget"
This reverts commit 9c100c95c48e1c32ee07b83cd158b47ec4678a5d, reversing
changes made to 6bed1be300ca9a9e7347657ac2f095cea6068897.

The reason for doing this is that ipywidgets is still moving much too
fast for Matplotlib to keep up with.  The original work done in #5754
was done against ipywidgets v4, as of now the final touches are being
put on v7.

We are reverting back to our old inject-into-the-DOM method from
`nbagg`, which is what is used by `%matplotlib notebook`.

For embedding as a 'proper' widget use
ipympl (github.com/matplotlib/jupyter-wigets) which will release on a
schedule that matches the upstream jupyter ecosystem.

closes #7695