Replace use of renderer._uid by weakref.
This avoids the need for a note regarding issues with renderer classes
that do not inherit from RendererBase.

Remember that the goal of _uid was to prevent two renderers from sharing
the same id() even though they were actually different (if one has been
GC'd and another is created at the same address).  Maintaining a weakref
to the renderer works as well, as weakref equality is thusly defined:

    If the referents are still alive, two references have the same
    equality relationship as their referents (regardless of the
    callback). If either referent has been deleted, the references are
    equal only if the reference objects are the same object.