MNT: ensure no draws are pending in paintEvent
This prevents flickering during window resize.  The issue is that the
resize event triggers a paint and we have marked the figure as stale
and requiring a re-draw.  Sometimes the paint from the resize will
process before the single-shot to trigger __draw_idle_agg and there
will be a size-mismatch between the QImage and the buffer we are
trying to fill it with which results in a visible flicker.