Fix tkagg animation error at exit.
Apparently, with a small enough animation interval, tk does not have
the time to fully proceed through FigureManagerTkAgg.show before the
animation takes over.  Removing the call to self.window.update (which
had been added back in 2004 (6c96389), but seems unnecessary (including
on Windows)) fixes the issue.

Also update the animation example so that it can serve as manual test
for the issue; changed the rate at which the data moves to keep it
reasonable, and switched to None (= count()) as index iterable to avoid
a jump every time we reach the end of the indices.