Revert "Merge pull request #8539 from Tuan333/tuan"
This reverts commit 3e2e31e7186d08e9a163ec96a5c154ef339763b5, reversing
changes made to 81495363d283809021b5ae4fc87bbd5edbbb3f6c.

This has some un-intended consequences (see #9261).  This exposes a
bigger issue about drawing on to non-linear axes which is if we
transform only the vertices or the full paths.  Which the user wants
is dependent on context so we need to expose this at the user API in a
nice way, but is a bigger change than we can accommodate for 2.1.

Reverting this change to return to the (slightly) broken state which
has been the satus-quo 7 years.