Update docstrings and allow infinite counts for plot_{surface,wireframe}.
Docstrings: Emphasize the downsampling.  Clarify when the defaults for
rstride and cstride apply.

Passing an infinite rcount and ccount is the easiest way to ensure the
data does not get downsampled (which I still think is a misfeature, but
that's a battle for another day), but would result in a stride of zero;
instead, ensure that the stride is at least 1.