Ask Appveyor to ignore certain branches.
This should prevent the automatic backports to be tested both when the
branch is pushed by the backport bot, and when the PR is opened.

Hopefully the PR will be still be tested (that's unclear from the
AppVeyor docs).

This was discussed in
https://github.com/matplotlib/matplotlib/pull/9283#issuecomment-334530421

And likely need to be also applied to other CI, but Appveyor is the most
annoying.