Update images.py
Instead of using the `print()`, directly call the object, which results in a reproducible line of code, as opposed to hard-coding the result into the tutorial text which isn't reproducible. In other words, in IPython this correction results to an output `Out[5]:` which actually corresponds to the input `In[5]: img`. In its current state, an output `Out[5]` is generated by the `print()`-command which isn't visible in the tutorial; only the assignment `In[5]: img = mpimg.imread('../../doc/_static/stinkbug.png')` is. Hence the relation in the tutorial between `In[5]:` and `Out[5]` is simply incorrect and not reproducible.