FIX: provide __ne__ implementation for transforms in py2
Closes #9455

The issue is that due to `!=` not defaulting to `not a == b` in
python2 (unlike python3) an internal check was failing and the y
ticks (aka the radial ticks) were being regenerated on every draw
and the alpha settings did not properly propagate.

This fix is a band-aid to make the behavior the same on all versions
of python.