In LogTransform, clip after log, not before.
When clipping before taking the log, the clip value in the output array
cannot be beyond log(epsilon) ~ -300.  When clipping after taking the
log, the clip value can be much further (see comments regarding choice
of value).

This allows setting the default nonpos mode back to "clip".

Some test images had to be updated as this effectively moves the
position of the "clip point", slightly changing the slope of the line
that goes to that point and thus the rasterization (by a tiny amount).