Rework backend loading.
This is preliminary work towards fixing the runtime detection of default
backend problem (the first step being to make it easier to check whether
a backend *can* be loaded).  No publically visible API is changed.

Backend loading now defines default versions of `backend_version`,
`draw_if_interactive`, and `show` using the same inheritance strategy as
builtin backends do.

For non-interactive backends (which don't override `mainloop`), restore
the default implementation of `show()` that prints a warning when run
from a console (the backend refactor accidentally removed this error
message as it provided a silent default `show()`).

The `_Backend` class had to be moved to the end of the module as
`FigureManagerBase` needs to be defined first.  The `ShowBase` class had
to be moved even after that as it depends on the `_Backend` class.