Check dependencies at runtime as declared in setup.py.
The goal is to list the requirements exactly once in code (in
setupext.py) and once in the docs (in INSTALL.rst).

Instead of manually importing and checking that (python) dependencies
are installed with the correct versions, rely on pkg_resources, which
allows automatically keeping things in sync with setupext, as well as
correctly check e.g. the complex version requirements on pyparsing.

Remove the overly complex way to specify the minimum numpy version.

Add comments reminding to update INSTALL.rst when the requirements are
changed in setupext.py; reorder dependencies in INSTALL.rst in a more
logical order.

Remove unneeded reference to MATLAB being a registered trademark.

Note that if you are running Matplotlib from source by manipulating
PYTHONPATH, then it will not appear to pkg_resources.  In this case,
there is no check for other dependencies (you are assumed to be
responsible enough for that).  Conversely, Matplotlib will fail to
import if it is installed into site-packages but some dependency (six,
etc.) is only available in PYTHONPATH instead of being correctly
installed.  I think this is acceptable (or we could just completely drop
the import checks and let bad installs fail with normal ImportErrors).