Rename some helper classes in dviread.
Currently, the existence of `matplotlib.dviread.Text` prevents one to
write `\`~.Text\`` in the docs when referring to `matplotlib.text.Text`
because sphinx does not know which of the two classes we are referring
to.

Rename `matplotlib.dviread.Text` to `_Text`, as there isn't (AFAICT) any
reason to use it externally.  Also renamed `_Page` and `_Box` for
consistency.

It turns out that the class deprecation machinery does not work for
deprecating namedtuples, but attaching the deprecation warning to
`__new__` instead of `__init__` fixes that.