dates._from_ordinalf: Use normal rounding for microseconds
Whenever the numeric input is small enough to provide actual microsecond
precision, we don't want to always round down, but instead to the
nearest microsecond.

The difference becomes visible when the AutoDateLocator picks the
MicrosecondLocator, and the format string for the tick labels is
configured to show microseconds. Support for both was added in
Matplotlib 2.0.