dates: select microsecond precision in AutoDateFormatter
DateLocator: viewlim_to_dt and datalim_to_dt should always get the most
accurate datetimes we can offer, so turn off the newly introduced
interval rounding for them.

AutoDateFormatter: We need to decide whether the tick labels should be
rounded when formatting them or not.  Rounding is most likely always
desired, except when exact microseconds are needed.

Inclusion of microsecond resolution in the strftime format string is
used as an indication that the user really wants to see microseconds and
they are relevant, so simply do not round them in that case.