dates: add warning for usec precision with high years
The warning threshold of is based on the assumption that the user really
wants full microsecond resolution in the plot, and expects the plot
frame to come out nicely (no jagged corners). Besides being able to
identify accurate tick labels/positions, it requires a small "accuracy
margin" for the drawing operations.

If we assume a desired accuracy of 50ns in the underlying numerical time
data, we need to go back to at least the year 27. It cannot hurt to
start warning about the issue slightly earlier, hence the treshold is
set to 20.