FIX: prevent recursive draws with plt.ion not in IPython with qt5agg
In 9b8a944db954c8240e14413e517e1b781881b80b the latch logic was moved
above the call to `self.draw` to de-latch the call to
`__draw_idle_agg` in `paintEvent` to protect against `blit` calls
during a draw.

However, this now fails to properly latch the base draw so with
`plt.ion` (not in IPython) the stale callback will cause a (possibly
infinite) recursion in the draw due to stale being set as part of the
draw call (which will re-trigger the draw).

This adds a second flag to track if we are currently rendering and
bail.

closes #10140