Deprecate the backend.qt{4,5} rcParams.
1) For the rcParams rewrite (MEP32), they may make it tricky to
transform rcParams into a nested dictionary type of object, as "backend"
is already the name of another (more important) rcParam.  This specific
issue could be solved with deprecation + renaming, but also...

2) The use of these rcParams can easily be replaced by a) first
importing one of the bindings (qt_compat will ensure that the already
imported binding gets used, as importing multiple bindings in the same
process is a bad idea anyways), or b) setting the QT_API environment
variable appropriately.