Deprecate truncating saved unsized anims to 100 frames.
To test that the deprecation is working, uncomment the `ani.save(...)`
line in the simple_anim.py example and modify the anim's save_count.

Animations that happened to generate exactly 100 frames without
truncation will incorrectly trigger a warning; not sure we can do much
about it (because after the 100th frame is generated there is no way to
know whether another frame is coming except by actually calling next()
on the generator, which we do not want to do).