Make the horizontal bar appear in AnchoredArtists example.
The `AnchoredSizeBar` should, according to its comment string 

> Draw a horizontal bar with the size in data coordinate of the given
        axes. A label will be drawn underneath (center-aligned).

This bar is not drawn, because from matplotlib 2.0 on, patches do not have any edge assigned to them any more. While the facecolor is set to `"None"`, it is really the edgecolor, which should be set to some color (e.g. black). This PR fixes this such that a bar appears in the plot.