Hide the backend.qt4/5 rcparam deprecation warning in test suite.
Otherwise, the test suite is spammed by warnings.  Note that pytest
explicitly disregards warnings filters (pytest issue 2430) so we can't
just set a filter when resetting the rcparams are reset in the test
suite.

The changes in matplotlib.testing are just there to allow it to be
imported relatively early in the matplotlib import process.