Make FigureCanvasWx ctor signature similar to FigureCanvasBase.
Currently, `FigureCanvasWx` has a signature of `(parent, id, figure)`,
unlike the Canvas classes of other backends, which only take `figure` as
argument.  Change `FigureCanvasWx` to be more similar to other backends.

A similar fix will later be applied to `FigureManagerWx` to bring it in
line with other Manager classes.

The ultimate goal is to get rid of `new_figure_manager` and
`new_figure_manager_given_figure` in the backend definitions, which can
be replaced by `return FigureManager(FigureCanvas(Figure(...)), num)` and
`return FigureManager(FigureCanvas(fig, num))` respectively.

Also deprecate the aliases `FigureCanvasWx.frame` and `.tb` (which are
named `.window` and `.toolbar` in all other backends).