Work towards removing reuse-of-axes-on-collision.
Currently, Matplotlib reuses axes when add_axes() is called a second
time with the same arguments.  This behavior is deprecated since 2.1.

However we forgot to deprecate the same behavior in gca(), so we can't
remove that behavior yet.

Also cleanup docstrings of Stack class.  Also,
process_projection_requirements cannot modify the outer kwargs (because
`**kwargs` is always a copy), so remove the incorrect note regarding the
need for copies.