Proof of concept for deprecation of global non-callables.
Triggered by the recent discussion on deprecating matplotlib.verbose.

This ensures that a deprecation warning is triggered just by trying to
access matplotlib.verbose (or importing it).  The implementation idea is
based on njsmith's metamodule, but without the hacks needed to support
Py<3.5.

Py3.7+ will provide instance-`__getattr__` on all classes by default,
making the `_DeprecatorModuleType` patching unnecessary (but the warning
emitter would still need to be implemented, so there is actually fairly
limited improvement from that).