Adjusted matplotlib.widgets.Slider to have optional vertical orientation.
The slider widget now takes an optional argument (orientation) with
values 'h' or 'v, which defaults to 'h'.

Argument checking is in keeping with the existing code, and the actual
changes to the source are minimal, replacing hspans, hlines and xdata with a
if switch to vspans and vlines and ydata.