Update INSTALL.rst.
Mostly: update the docs re: Windows (on 3.5 setting up the build
environment is much easier, coming down to just "install MSVC" -- we may
as well just link to an official doc page rather than a blog).

The conda recipe is not hosted in the repo anymore, update that line in
the docs.

Minor rewordings here and there.