Fix overflow when resizing path-to-string buffer.
The int version of the buffer size was not updated when the buffer was
resized. It's there to prevent a signed/unsigned comparison warning, but
it's simpler just to cast the other side of the comparison. There's no
problem with the signed-to-unsigned cast since we already know that the
result is positive due to the previous check.

Fixes #10889.