Add PEP8-compliant aliases to transAxes, transData, etc.
ax.transAxes / transData / etc. are perhaps one of the most user-facing
parts of mpl that are not pep8 compliant.  Provide some PEP8-compliant
aliases for those (like myself) who are excessively (and unduly?)
bothered by that.

Note that (quite unusually for myself :-)) I am **not** proposing to
deprecate the camelCase names.