Make Marker.get_snap_threshold() always return a scalar.
... instead of possibly None.

The snap threshold indicates sizes of markers above which the marker
would be drawn as snapped to integer pixels, avoiding antialiasing
issues.  Some markers used to set it to None, which meant "snap only if
the path is fully made of vertical or horizontal segments" (see
GraphicsContextBase.set_snap for documentation and path_converters.h,
line 471+ for implementation).  But the marker path is actually known
and fixed, so that case can simply be replaced by setting the snap
threshold to either 0 or inf depending on whether the marker path
satisfies the above condition.

The only case where the snapping is changed is if one passed in a
*custom* path with only vertical or horizontal segments as marker
-- or used the `(numsides, style, angle)` format to achieve that
(which basically means a line, a square, or a plus sign, all of which
are available as standard markers).  I'd argue that the additional
complexity for what is a very rare case is not worth it.

The deprecation is only needed for someone who'd have overridden
get_snap_threshold in a Marker subclass and somehow installed such a
marker (which is pretty close to API abuse, but heh).