When drawing markers, don't set the GraphicsContext alpha.
Setting the alpha value on the GraphicsContext forcefully applies it to
both the edgecolor and the facecolor, *except* that a facecolor of
`None` ignores the alpha.

Instead of requiring this behavior (which needlessly complicates the
implementation of new rendering backends), the alpha value (if any) can
directly be preapplied to the edgecolor and facecolor, the edgecolor can
be applied using gc.set_foreground, and the alpha value never set on the
GraphicsContext.