Group figure.subplot.* rc to a single rcParam.
The idea is to make it easier to restore a figure's subplotparams to the
rc-provided defaults.  Specifically, this can now be done with

    fig.subplots_adjust(**rcParams["figure.subplot"])

or

    fig.subplots_adjust(**dict(rcParams["figure.subplot"], left=...))

if some values need to be overridden.

Also make Figure.clf() restore the subplotsparams to these rc-provided
defaults (that is the original motivation for the change).

Note that this PR runs into a limitation of the rcParams API: it would
be nicer if the validators took the rcParams as parameters (e.g.
implicitly by being methods of the RcParams class), so that they can
actually depend on another key in the same instance.