Automagically set the stacklevel on warnings.
There are many places in Matplotlib where it is impossible to set a
static stacklevel on warnings that works in all cases, because a same
function may either be called directly or via some wrapper (e.g.
pyplot).

Instead, compute the stacklevel by walking the stack.  Given that
warnings refer to conditions that should, well, be avoided, I believe
we don't need to worry too much about the runtime cost.

As an example, use this mechanism for the "ambiguous second argument to
plot" warning.  Now both
```
plt.gca().plot("x", "y", data={"x": 1, "y": 2})
```
and
```
plt.plot("x", "y", data={"x": 1, "y": 2})
```
emit a warning that refers to the relevant line in the user source,
whereas previously the second snippet would refer to the pyplot wrapper,
which is irrelevant to the user.

Note that this only works from source, not from IPython, as the latter
uses `exec` and AFAICT there is no value of stacklevel that correctly
refers to the string being exec'd.

Of course, the idea would be to ultimately use this throughout the
codebase.