Add private API retrieving the current event loop and backend GUI info.
Work towards implementation of backend switching.

Note that the API is kept private for now as the lack of extensibility
is a bit unsatisfying; perhaps we'll figure out a better way to do it
if other custom implementers (e.g., code editors?) get interested in
looking into it.

The `required_event_loop` variable could be made private for now, or
not...