ENH: backend switching.
See changes documented in the API changes file.

I inlined pylab_setup into switch_backend (and deprecated the old
version of pylab_setup) because otherwise the typical call stack would
be `use()` -> `switch_backend()` -> `pylab_setup()`, which is a bit of
a mess; at least we can get rid of one of the layers.