Update blocking_input.py
ginput doesn´t allow to disable buttons.
With this change you can assign an invalid button to the three mouse buttons and leave the left click empty to zoom or pan.

Updated the documentation associated with ginput, regarding the changes applied to blocking_input to allow for pan and zoom

(PR cleanup by the committer.)