Don't draw axis (spines, ticks, labels) twice when using parasite axes.
We can just defer to the superclass' draw (axes._base.Axes.draw)
which will correctly draw everything in get_children();
axislines.Axes.get_children already adds the axislines and gridlines to
the superclass' get_children().

Likewise, get_tightbbox() doesn't need to be overridden because the
superclass' get_tightbbox by default uses
get_default_bbox_extra_artists, which fetches all visible children.
(This point is not needed for the bugfix, but we may as well...)

Don't bother using remove_text in the new figures, because that doesn't
remove the labels anyways (and the tests are also about label placement).