Don't assume cwd in test_ipynb.
The Matplotlib test suite can be run from outside of the root git
directory, e.g. with

    pytest --pyargs matplotlib.tests.test_backend_nbagg

(mplcairo explicitly relies on that to reuse the Matplotlib test suite).

To ensure this, find the example notebook relative to the path of the
test module, rather than relative to the current working directory.