Change {FigureCanvasAgg,RendererAgg}.buffer_rgba to return a memoryview.
The `buffer_rgba` method now allows direct access to the renderer's
underlying buffer (as a `(m, n, 4)`-shape memoryview) rather than
copying the data to a new bytestring.  This is consistent with the
behavior on Py2, where a buffer object was returned.

While this is technically a backwards-incompatible change, memoryviews
are in fact quite compatible with bytes for most uses, and I'd argue
that the bigger compatibility break was the change from no-copy in Py2
to copy in Py3 (after all that's the main point of the method...).