Allow to not draw the labels on pie chart.
for consistency with other API it is relatively easy and desirable to be
able to pass the `label` and call `legend()` after the fact; which will
display the legend in a box in one of the corner.

When doing so, having the label on each piece of pie does not make much
sens (and may clobber the pie chart). Thus allow to use ``None`` as a
value for ``labeldistance`` as a way to not draw them.