Resolve backend in rcParams.__getitem__("backend").
... to avoid leaking the non-string sentinel.

Changes to `backends/__init__` and `_backend_selection()` were necessary
to avoid import-time cycles.

The fallback order in `backend_fallback` slightly changed in that the
FooCairo backends now fallback to WxAgg instead of Wx when a wx event
loop is active, but I wouldn't worry too much about it anyways given
that the Wx backend is deprecated and the correct fallback would be
WxCairo to start with.