don't share axes if share_all is False
If share_all is False, we should not share any axes. As it is written, in a 2x2 grid, it will share the row axis by the second cell, the column axis by the third, and both aces by the fourth. It looks like this code was copied from class GridAxes; however, that class has share_x and share_y options (along with share_all). For some reason, it is the share_x (share_y) code path that has been copied here.